+++ lib/features/select_your_plan_hours/presentation/components/choice_chips_list_view.dart
@@ -5,7 +5,6 @@ import 'package:test1/core/constants/constants.dart';
 import 'nationality_chip.dart';
 
 class NationalityChipsListView extends StatefulWidget {
-
   const NationalityChipsListView({super.key});
 
   @override
@@ -24,7 +23,8 @@ class _NationalityChipsListViewState extends State<NationalityChipsListView> {
         scrollDirection: Axis.horizontal,
         itemBuilder: (context, index) {
           return Padding(
-            padding:  EdgeInsets.only(left: 5.w,right: index==0?AppConstants.k30ViewPadding:0),
+            padding: EdgeInsets.only(
+                left: 5.w, right: index == 0 ? AppConstants.k30ViewPadding : 0),
             child: NationalityChip(
               isSelected: currentIndex == index,
               onTap: () {
